requirements.txt
# annotated-types==0.6.0
# anyio==3.7.1
bcrypt==4.1.2
# certifi==2023.11.17
# cffi==1.16.0
# click==8.1.7
# colorama==0.4.6
cryptography==41.0.7
# ecdsa==0.18.0
# exceptiongroup==1.2.0
fastapi==0.108.0
# h11==0.14.0
# httpcore==1.0.2
httpx==0.26.0
# idna==3.6
# iniconfig==2.0.0
Jinja2==3.1.2
# MarkupSafe==2.1.3
mysql-connector-python==8.3.0
# packaging==23.2
passlib==1.7.4
# pluggy==1.3.0
# protobuf==4.21.12
# pyasn1==0.5.1
# pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.1.0
# pydantic_core==2.14.6
pytest==7.4.3
# python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
# rsa==4.9
# six==1.16.0
# sniffio==1.3.0
# starlette==0.32.0.post1
# tomli==2.0.1
# typing_extensions==4.9.0
uvicorn==0.25.0
